% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-id.R
\name{aggregate_id}
\alias{aggregate_id}
\title{Aggregate the id axis based on an id-invariant variable}
\usage{
aggregate_id(dt, val, along, ..., FUN = mean)
}
\arguments{
\item{dt}{the data cube}

\item{val}{the value column to aggregate}

\item{along}{the id-invariant variable to aggregate upon}

\item{...}{additional operations on the \code{along} variable}

\item{FUN}{a function to aggregate by}
}
\value{
a cube with id being aggregated
}
\description{
Aggregate the id axis based on an id-invariant variable
}
\examples{
\dontrun{
library(stringr)
water_small_cube <- water_small \%>\%
  mutate(river_name = str_extract(name, "^[^\\\\@]+ "),
         river_name = as.factor(str_trim(river_name)),
         river_name = tolower(word(river_name, 1, 2, sep = " "))) \%>\%
  as_tsibble(key = c(parameter, station), index = date) \%>\%
  as_cube(param = parameter)
aggregate_id(water_small_cube, value, long, ~ggplot2::cut_interval(.x, 3))
aggregate_id(water_small_cube, val = value, along = river_name)
}
}
