% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate-param.R
\name{aggregate_param}
\alias{aggregate_param}
\title{Aggregate the id axis based on an id-invariant variable}
\usage{
aggregate_param(dt, weight)
}
\arguments{
\item{dt}{the data cube}

\item{weight}{a weighting matrix for multivariate parameters}
}
\value{
a cube with parameters being aggregated
}
\description{
Aggregate the id axis based on an id-invariant variable
}
\examples{
\dontrun{
library(stringr)
water_small_cube <- water_small \%>\%
  mutate(river_name = str_extract(name, "^[^\\\\@]+ "),
         river_name = as.factor(str_trim(river_name)),
         river_name = tolower(word(river_name, 1, 2, sep = " "))) \%>\%
  as_tsibble(key = c(parameter, station), index = date) \%>\%
  as_cube(param = parameter)
v <- c(0.1, 0.2, 0.25, 0.3, 0.1, 0.05)
aggregate_param(water_small_cube, v)
}
}
