% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate the time axis}
\usage{
aggregate_time(dt, period, summarised_by)
}
\arguments{
\item{dt}{the data cube}

\item{period}{the period to be aggregated into}

\item{summarised_by}{the summary function applied to aggregation}
}
\description{
Aggregate the time axis
}
