% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{as_cube}
\alias{as_cube}
\title{casting and coercion}
\usage{
as_cube(x, key = NULL, index, param = NULL, ...)
}
\description{
casting and coercion
}
\examples{
\dontrun{
water <- water_raw \%>\%
as_tsibble(key = c(parameter, station), index = time, regular = FALSE) \%>\%
as_cube(param = parameter)
pedestrian \%>\%
tidyr::pivot_longer(cols = Count, names_to = "count", values_to = "val") \%>\%
as_cube(param = count)
}
}
