% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attrs.R
\name{key_vars.cubble_df}
\alias{key_vars.cubble_df}
\alias{key.cubble_df}
\alias{key_data.cubble_df}
\alias{coords}
\alias{spatial}
\alias{spatial.spatial_cubble_df}
\alias{spatial.temporal_cubble_df}
\alias{index}
\alias{index_var}
\title{Extract cubble attributes}
\usage{
\method{key_vars}{cubble_df}(x)

\method{key}{cubble_df}(x)

\method{key_data}{cubble_df}(.data)

coords(data)

spatial(data)

\method{spatial}{spatial_cubble_df}(data)

\method{spatial}{temporal_cubble_df}(data)

index(data)

index_var(data)
}
\arguments{
\item{x, .data, data}{a cubble object}
}
\description{
Extract cubble attributes
}
\examples{
library(tsibble)
key(climate_mel)
key_vars(climate_mel)
key_data(climate_mel)
cubble::index(climate_mel)
cubble::index_var(climate_mel)
coords(climate_mel)
spatial(climate_mel)
}
