% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubbe-s3.R
\name{form}
\alias{form}
\alias{meta}
\alias{group_vars}
\alias{groups}
\title{get attributes for a cubble object}
\usage{
form(data)

meta(data)

group_vars(data)

groups(data)
}
\arguments{
\item{data}{an cubble object}
}
\description{
get attributes for a cubble object
}
