% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aus_climate}
\alias{aus_climate}
\title{Australia climate data}
\format{
A cubble object with 5 station-related variables and 3 time-related variables (PRCP, TMAX, and TMIN), nested in the \code{ts} column
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation (WMO) station number}
\item{ts}{a list-column that nests all the time-wise measures: date, prcp, tmax, and tmin}
}
}
\usage{
aus_climate
}
\description{
One year worth (2020) of climate data recorded  in 72 stations across Australia.
The data is computed from National Oceanic and Atmospheric Administration (NOAA) using the \code{rnoaa} package.
}
\examples{
library(rmapshaper)
library(ggplot2)
state_map <- ms_simplify(ozmaps::abs_ste, keep = 2e-3)
plot_map(state_map) +
  geom_point(data = aus_climate, aes(x = long, y = lat))
}
\keyword{datasets}
