% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\title{extract bbox of the coordinate with a buffer}
\usage{
bbox(dt, lat, long, buffer_ratio = 0.01)
}
\arguments{
\item{dt}{the data object}

\item{lat}{the latitude column}

\item{long}{the longitude column}

\item{buffer_ratio}{the percentage to add/deduct to the max/min lat/long when querying the map}
}
\description{
extract bbox of the coordinate with a buffer
}
\examples{
aus_map <- get_map(location = bbox(nested), source = "osm")
ggmap(aus_map) +
   geom_point(data = nested, aes(x = long, y = lat))
}
