% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R
\name{is_cubble}
\alias{is_cubble}
\alias{is_cubble_spatial}
\alias{is_cubble_temporal}
\alias{is_sf}
\alias{is_tsibble}
\title{Predicate functions on the object class}
\usage{
is_cubble(data)

is_cubble_spatial(data)

is_cubble_temporal(data)

is_sf(data)

is_tsibble(data)
}
\arguments{
\item{data}{an object to test for the class}
}
\value{
a logical value of TRUE/FALSE
}
\description{
Predicate functions on the object class
}
\examples{
is_cubble(stations)
is_cubble(meteo)
is_cubble(climate_flat)
is_cubble(climate_mel)
is_cubble(climate_aus)
is_cubble_spatial(climate_aus)
is_cubble_temporal(climate_aus)
}
