% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R
\name{is_cubble}
\alias{is_cubble}
\alias{is_cubble_spatial}
\alias{is_cubble_temporal}
\title{Check if the object is a cubble or its subclass}
\usage{
is_cubble(data)

is_cubble_spatial(data)

is_cubble_temporal(data)
}
\arguments{
\item{data}{the object to test}
}
\value{
a TRUE/FALSE predicate
}
\description{
Check if the object is a cubble or its subclass
}
\examples{
is_cubble(stations)
is_cubble(meteo)
is_cubble(climate_flat)
is_cubble(climate_mel)
is_cubble(climate_aus)
is_cubble_spatial(climate_aus)
is_cubble_temporal(climate_aus)
}
