% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_key.R
\name{check_key}
\alias{check_key}
\title{Check key matching from multiple data sources}
\usage{
check_key(spatial, temporal, by = NULL)
}
\arguments{
\item{spatial}{a tibble}

\item{temporal}{a tibble}

\item{by}{in the syntax of \code{by} in \code{left_join()} to specify the key
when they have different names in the spatial and temporal data}
}
\value{
a list with three elements: 1) paired: a tibble of paired ID from
spatial and temporal data, 2) potential_pairs: a tibble of pairs that could
potentially match from both datasets, 3) others: other key values that can't
be matched in a list: others$temporal and others$spatial
}
\description{
When creating a cubble from two sources (spatial and temporal data) with
\code{as_cubble(list(spatial = ..., temporal = ...), ...)}, \code{as_cubble()} will
informed users about potential disagreement of the key values in the two datasets
(some sites appear in one table but not the other). This function summarises
the key values into those match, potentially can be matched, and can't be matched.
}
\examples{
three_sites <-  stations \%>\% head(3)
five_ts <- climate \%>\% filter(month(date) == 1)
check_key(spatial = three_sites, temporal = five_ts)

five_ts2 <- five_ts \%>\% rename(station = id)
check_key(spatial = three_sites, temporal = five_ts2, by = c("id" = "station"))
}
