% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose}
\alias{choose}
\title{Choose which sheet to work on}
\usage{
choose(df, sheet)
}
\arguments{
\item{df}{a cubble object}

\item{sheet}{one of the sheet in the cubble object}
}
\description{
Choose which sheet to work on
}
