% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate}
\alias{climate}
\title{Australian climate data in 2020}
\format{
A tsibble with 87,639 rows and 13 columns
\describe{
\item{date}{date of the record}
\item{station}{weather stations recorded. See \code{station} dataset for metadata of each station}
\item{prcp:tavg}{variables recorded in each station. See \url{https://www1.ncdc.noaa.gov/pub/data/ghcn/daily/readme.txt} for detailed data description)}
}
}
\usage{
climate
}
\description{
Australian climate data in 2020
}
\examples{
library(tidyr)
climate \%>\% pivot_longer(prcp:tavg, names_to = "param", values_to = "value")
}
\keyword{datasets}
