% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_large}
\alias{climate_large}
\title{Australia climate data (large)}
\format{
A cubble object with 5 spatio-related variables and 11 time-related variables, nested in the \code{ts} column
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation station number, if has one}
\item{ts}{a list-column that nests all the time-wise measures:
date, prcp, tmax, tmin, dapr, datn, datx, dwpr, mdpr, mdtn, mdtx, tavg}
}
}
\usage{
climate_large
}
\description{
One year worth (2020) of climate data recorded  in 478 stations across Australia.
The data is computed from the two raw data \link{climate} and \link{station}
obtained from National Oceanic and Atmospheric Administration (NOAA) using the \code{rnoaa} package.
}
\seealso{
climate station
}
\keyword{datasets}
