% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_missing}
\alias{climate_missing}
\title{Climate data with missing value}
\format{
a cubble object with 50 rows and 7 columns
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation (WMO) station number}
\item{ts}{a list-column that nests all the time-wise measures: date, prcp, tmax, and tmin}
}
}
\usage{
climate_missing
}
\description{
Climate data with missing value
}
\keyword{datasets}
