% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_small}
\alias{climate_small}
\title{climate_small}
\format{
a climate dataset with 55 stations recorded across Australia
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{ts}{a list-column with variables \code{date}, \code{prcp}, \code{tmax} and \code{tmin} nested}
}
}
\usage{
climate_small
}
\description{
climate_small
}
\keyword{datasets}
