% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attr.R
\name{coords}
\alias{coords}
\title{extract the coords attribute from a cubble}
\usage{
coords(data)
}
\arguments{
\item{data}{a cubble object}
}
\description{
extract the coords attribute from a cubble
}
\examples{
coords(climate_mel)
}
