% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R, R/cubble-s3.R
\name{is_cubble}
\alias{is_cubble}
\alias{cubble}
\title{Create a cubble object}
\usage{
is_cubble(data)

cubble(..., key, index, coords)
}
\arguments{
\item{data}{object to be converted into an object of class \code{cubble_df}}

\item{...}{a list object to create new cubble}

\item{key}{the variable that identifies the spatial location.}

\item{index}{the time identifier}

\item{coords}{the coordinates that characterise the spatial dimension}
}
\value{
a TRUE/FALSE predicate

a cubble object
}
\description{
Create a cubble object
}
