% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-package.R
\docType{package}
\name{cubble-package}
\alias{cubble-package}
\alias{_PACKAGE}
\title{cubble: A Relational Data Structure for Spatiotemperal Data}
\description{
A spatiotemperal data object in a relational data structure to separate the recording of time variant/ invariant variables.
}
\author{
\strong{Maintainer}: Sherry Zhang \email{huize.zhang@monash.edu} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID})

}
\keyword{internal}
