% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R
\name{print.cubble_df}
\alias{print.cubble_df}
\alias{tbl_sum.spatial_cubble_df}
\alias{tbl_sum.temporal_cubble_df}
\title{Print methods}
\usage{
\method{print}{cubble_df}(x, width = NULL, ...)

\method{tbl_sum}{spatial_cubble_df}(x)

\method{tbl_sum}{temporal_cubble_df}(x)
}
\arguments{
\item{x}{any \R object (conceptually); typically numeric.}

\item{width}{\code{default} method: the \emph{minimum} field width or
    \code{NULL} or \code{0} for no restriction.

    \code{AsIs} method: the \emph{maximum} field width for non-character
    objects.  \code{NULL} corresponds to the default \code{12}.
  }

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print methods
}
\examples{
climate_mel # a nested/spatial cubble
face_temporal(climate_mel) # a long/temporal cubble
}
