% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Incorporate group-wise metadata into a long form cubble}
\usage{
migrate(data, ...)
}
\arguments{
\item{data}{a long form tibble object}

\item{...}{variables in metadata to included into the long form}
}
\description{
Incorporate group-wise metadata into a long form cubble
}
\details{
Some spatio-temporal visualisations, i.e. a glyph map, require both group-wise variables
and time-wise variables to make the plot. While data manipulation can be easier when separating
those two types of variables, the last step before plotting the data is to transfer the useful
group-wise variables into the long form. \code{migrate()} is a shortcut for transferring group-wise
variables from cubble's metadata. For transfering other group-wise variables, use \code{left_join}.
}
\examples{
climate_small \%>\% zoom() \%>\% migrate(lat, long)
}
