% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relational.R
\name{cubble}
\alias{cubble}
\alias{new_cubble}
\alias{as_cubble}
\alias{is.tbl_cb}
\title{Create a cubble object}
\usage{
cubble(main, item, by = NULL)

new_cubble(main, item, by)

as_cubble()

is.tbl_cb(x)
}
\description{
Create a cubble object
}
\examples{
df <- cubble(climate, station, by = c("station" = "id"))
}
