% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-join.R
\name{cubble_join}
\alias{cubble_join}
\title{Join the result to the cubble object}
\usage{
cubble_join(single_cb)
}
\arguments{
\item{single_cb}{a single_cb object}
}
\description{
Join the result to the cubble object
}
