% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/face-spatial.R
\name{face_spatial}
\alias{face_spatial}
\alias{face_spatial.cubble_df}
\alias{face_spatial.spatial_cubble_df}
\alias{face_spatial.temporal_cubble_df}
\title{Switch a cubble object into the nested form}
\usage{
face_spatial(data)

\method{face_spatial}{cubble_df}(data)

\method{face_spatial}{spatial_cubble_df}(data)

\method{face_spatial}{temporal_cubble_df}(data)
}
\arguments{
\item{data}{a long cubble object}
}
\value{
a cubble object in the nested form
}
\description{
\code{face_spatial()} turns a long cubble back into a nest cubble and can be seen as
the inverse operation of \code{face_temporal()}.
The nested cubble identifies each row by \code{key} and is suitable
for operations whose output doesn't involve a time index.
}
\examples{
cb_long <- climate_mel \%>\% face_temporal()

# face_spatial and face_temporal are exact inverse
cb_back <- cb_long \%>\% face_spatial()
identical(climate_mel, cb_back)
}
