% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps.cubble_df}
\alias{fill_gaps.cubble_df}
\title{\code{tsibble} methods implemented in cubble}
\usage{
fill_gaps.cubble_df(.data, ..., .full = FALSE, .start = NULL, .end = NULL)
}
\arguments{
\item{.data, ..., .full, .start, .end}{see tsibble documentation}
}
\value{
a cubble object
}
\description{
See \code{\link[tsibble]{fill_gaps}}
}
