% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non-varying.R
\name{find_non_varying}
\alias{find_non_varying}
\title{Slice the data cube into lower dimension}
\usage{
find_non_varying(dt, key)
}
\arguments{
\item{dt}{the raw dataframe to slice, in long form - does it has to be long form?}

\item{key}{the dimension to slice}
}
\value{
a nested tibble/ tsibble - think through here
}
\description{
Slice the data cube into lower dimension
}
\examples{
\dontrun{
water_raw \%>\% find_non_varying(station)
water_raw \%>\% find_non_varying(param)
water_raw \%>\% find_non_varying(time)
}
}
