% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier.R
\name{get_centroid}
\alias{get_centroid}
\title{Find the centroid of cubble}
\usage{
get_centroid(data)
}
\arguments{
\item{data}{a cubble data object}
}
\value{
a cubble object
}
\description{
find the convex hull that wraps around the cluster, make it a polygon, find the centroid of the polygon and finally, extract the x and y coordinate of each centroid:
}
