% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{global}
\alias{global}
\title{Formulate or switch an object into a cubble list-column form}
\usage{
global(data, key)
}
\arguments{
\item{data}{the data to be converted into a cubble object}

\item{key}{the spatio identifier. Key can be automatically detected for a cubble object}
}
\description{
Create a cubble object or switch an existing cubble object into the list-column form
}
\details{
\itemize{
\item For an object that is not \code{cubble_df}, \code{global()} will initialise a cubble object upon
providing a \code{group} variable.
\item For an existing \code{cubble} in the long form, \code{global()}
switch the object back to the list-column form. This form is easier to computing
group-specific variables.s
}
}
\examples{
library(lubridate)
library(dplyr)

# create a cubble object from a tibble
climate_flat \%>\% global(station)

# switch to list-column form from the long form
climate_small \%>\%
  zoom() \%>\%
  filter(year(date) == 2020) \%>\%
  global()
}
\seealso{
Other cubble verbs include \code{\link{zoom}} and \code{\link{migrate}}
}
