% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attrs.R
\name{index}
\alias{index}
\alias{index_var}
\title{extract index related components in a cubble}
\usage{
index(data)

index_var(data)
}
\arguments{
\item{data}{a cubble object}
}
\description{
extract index related components in a cubble
}
\examples{
index(climate_mel)
index_var(climate_mel)
}
