% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{index_by.cubble_df}
\alias{index_by.cubble_df}
\title{tsibble related functions}
\usage{
\method{index_by}{cubble_df}(data, ...)
}
\arguments{
\item{data}{the data object to be operated on}

\item{...}{other arguments passed onto the function}
}
\description{
tsibble related functions
}
