% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R
\name{is_cubble}
\alias{is_cubble}
\title{Check if the object is a cubble or its subclass}
\usage{
is_cubble(data)
}
\arguments{
\item{data}{the object to test}
}
\value{
a TRUE/FALSE predicate
}
\description{
Check if the object is a cubble or its subclass
}
