% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-cb.R
\name{is_single_cubble}
\alias{is_single_cubble}
\title{Test whether an object is a single cubble}
\usage{
is_single_cubble(x)
}
\arguments{
\item{x}{a single cubble}
}
\description{
Test whether an object is a single cubble
}
