% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attrs.R
\name{key_vars.cubble_df}
\alias{key_vars.cubble_df}
\alias{key.cubble_df}
\alias{key_data.cubble_df}
\title{extract key related components in a cubble}
\usage{
\method{key_vars}{cubble_df}(x)

\method{key}{cubble_df}(x)

\method{key_data}{cubble_df}(.data)
}
\arguments{
\item{x, .data}{a cubble object}
}
\description{
extract key related components in a cubble
}
\examples{
library(tsibble)
key(climate_mel)
key_vars(climate_mel)
key_data(climate_mel)
}
