% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaves.R
\name{new_leaves}
\alias{new_leaves}
\title{Create and manipulate leaves attributes}
\usage{
new_leaves(data, key)
}
\arguments{
\item{data}{a flat data without any nesting structure}

\item{key}{the spatial identifier}
}
\description{
A leaves object records whether a variable in the cubble varies across time or not and
make it possible for a cubble to switch between the nested and long form.
This set of functions create and manipulate the leaves object.
}
\details{
\itemize{
\item \code{new_leaves()} creates a leaves object by supplying a flat data and the grouping variable
\item \code{is_leaves()} tests whether the object is a leaves object
\item \code{invariant()} and \code{variant()} extracts the time-variant and invariant variables from a leaves object
\item \code{tbl_sum()} provides a customised printing header for leaves
}

Notice that users may not need to these functions to directly manipulate the leaves.
To extract the leaves object from a cubble, use \code{leaves()}
}
