% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatio.R
\name{ll_join}
\alias{ll_join}
\title{Spatial join of two cubbles}
\usage{
ll_join(data1, data2)
}
\description{
An option to perform spatial join is to use \code{sf::st_join()}, however, this
requires 1) a data frame to be converted to an sf object based on long and lat,
2) use \code{sf::st_join()} for joining, and 3) convert the object back to a data
frame and drop the geometry for a sole operation. \code{ll_join()} uses package s2
to create a lnglat vector and perform the join directly on a cubble without
converting between object.
}
