% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R, R/tsibble.R
\name{make_spatial_sf}
\alias{make_spatial_sf}
\alias{make_temporal_tsibble}
\title{Promote the spatial/ temporal component in cubble to an sf/ tsibble object after creation}
\usage{
make_spatial_sf(data, ....)

make_temporal_tsibble(data, ...)
}
\arguments{
\item{data}{A cubble object}

\item{...}{not used}
}
\description{
Promote the spatial/ temporal component in cubble to an sf/ tsibble object after creation
}
\examples{
climate_mel \%>\% make_spatial_sf()
climate_mel \%>\% face_temporal() \%>\% make_temporal_tsibble()
}
