% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{make_spatial_sf}
\alias{make_spatial_sf}
\title{Update the spatial cubble to include the sf class}
\usage{
make_spatial_sf(x, sfc = NULL, crs, silent = FALSE)
}
\arguments{
\item{x}{object of class \code{spatial_cubble_df}}

\item{sfc}{object of class \code{sfc} (see package sf)}

\item{crs}{object of class \code{crs} (see package sf); if missing
'OGC:CRS84' is assumed (WGS84) and a message is emitted}

\item{silent}{logical; suppress message?}
}
\description{
add geometry list column to cubble_df object
}
\examples{
climate_mel \%>\% make_spatial_sf()
}
\seealso{
\link{make_temporal_tsibble}
}
