% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key-op.R
\name{match_key}
\alias{match_key}
\title{Matching}
\usage{
match_key(major, minor, n_slice = 1, dist_max = 10)
}
\arguments{
\item{major}{The major dataset to match, every key in the major dataset will have a match, unless filtered by \code{dist_max}}

\item{minor}{The dataset to match from}

\item{n_slice}{number of match to keep for each key in the major dataset}

\item{dist_max}{maximum distance allowed}
}
\value{
a tibble
}
\description{
Matching
}
