% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Incorporate meta data into the long form}
\usage{
migrate(data, ...)
}
\arguments{
\item{data}{a long form tibble object}

\item{...}{variables in metadata to included into the long form}
}
\description{
Incorporate meta data into the long form
}
\examples{
\dontrun{
out <- oz_zoom2 \%>\% migrate(lat, long)
# Currently, switch back to global model will remove those migrated variables
out \%>\% global()
}
}
