% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{add_missing_prct}
\alias{add_missing_prct}
\alias{add_missing_dscrb}
\title{Compute statistics for variable missingness}
\usage{
add_missing_prct(data, vars)

add_missing_dscrb(data, cutoff = c(0, 0.95))
}
\arguments{
\item{data}{a cubble object}

\item{vars}{the variable to compute missingness}

\item{cutoff}{the lower and upper threshold value, below or above which the variable
will be described as "almost no missing" and "almost all missing". ("some missing" for
if in between). Only used in \code{add_missing_dscrb()}}
}
\description{
Compute statistics for variable missingness
}
\examples{

\dontrun{
out <- oz_global2 \%>\% add_missing_prct(vars = c("prcp", "tmin", "tmax"))
out2 <- out \%>\% add_missing_dscrb()
}
}
