% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_nearby}
\alias{slice_nearby}
\alias{view_nearby}
\title{Autoplot for viewing the nearby points sampled}
\usage{
slice_nearby(data, ..., buffer = 1, n = 5)

view_nearby(data, origin, map = NULL)
}
\arguments{
\item{data}{a data object output from \code{slice_nearby()}}

\item{...}{conditions to filter for the selected data}

\item{buffer}{the buffer added to the bounding box for slicing}

\item{n}{the number of nearby points to sample}

\item{origin}{the original dataset supplied to \code{slice_nearby()}}

\item{map}{a map object as the background, default to the Australia country map}
}
\description{
Autoplot for viewing the nearby points sampled
}
\examples{
slice_nearby(climate_large, name == "Melbourne Airport") \%>\%
   view_nearby(origin = climate_large,
               map = ozmaps::abs_ste \%>\% dplyr::filter(NAME == "Victoria"))
}
