% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nested}
\alias{nested}
\title{A nested data}
\format{
A rowwise dataframe iwth 1,451 rows adn 7 columns
\describe{
\item{station}{station id}
\item{data}{a list-column that nests all the time-wise measures}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation station number, if has one}
}
}
\usage{
nested
}
\description{
A nested data
}
\keyword{datasets}
