% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-s3.R
\name{new_cubble}
\alias{new_cubble}
\title{Cubble constructor}
\usage{
new_cubble(data, key, index, coords, spatial, form, tsibble_attr = NULL)
}
\arguments{
\item{data}{data}

\item{key}{key}

\item{index}{key}

\item{coords}{key}

\item{spatial}{key}

\item{form}{key}

\item{tsibble_attr}{key}
}
\description{
Cubble constructor
}
