% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_climate}
\alias{oz_climate}
\title{oz_climate}
\format{
A rowwise dataframe iwth 1,451 rows and 7 columns
\describe{
\item{station}{station id}
\item{date}{date}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{prcp}{precipitation}
\item{tmax}{max temperature}
\item{tmin}{min temperature}
}
}
\usage{
oz_climate
}
\description{
oz_climate
}
\keyword{datasets}
