% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{ts_aggregate}
\alias{ts_aggregate}
\alias{normalising}
\title{Pipeline for index calculation}
\usage{
ts_aggregate(data, scale, var)

normalising(data, dist, gran = "month", var)
}
\arguments{
\item{data}{a cubble object}

\item{scale}{the time scale used for aggregation, in month}

\item{var}{the variable used to calculate the index}

\item{dist}{a distribution function, see details}

\item{gran}{the granulate for calculating the index, default to "month"}
}
\value{
a long cubble

a tibble object (end of a pipeline)
}
\description{
Pipeline for index calculation
}
