% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prcp_aus}
\alias{prcp_aus}
\title{Daily precipitation data from 2016 to 2020}
\format{
An object of class \code{cubble_df} (inherits from \code{rowwise_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 639 rows and 7 columns.
}
\usage{
prcp_aus
}
\description{
Daily precipitation data from 2016 to 2020
}
\keyword{datasets}
