% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep-tour.R
\name{prep_edges}
\alias{prep_edges}
\alias{prep_edges.cubble_df}
\alias{prep_data}
\alias{prep_data.cubble_df}
\title{A function to prepare edges data for tour display}
\usage{
prep_edges(data, edges_col, color_col)

\method{prep_edges}{cubble_df}(data, edges_col, color_col = NULL)

prep_data(data, cols)

\method{prep_data}{cubble_df}(data, cols = NULL)
}
\arguments{
\item{data}{a cubble object}

\item{edges_col}{the variable maps to edges colour}

\item{color_col}{the variable maps to point colour}

\item{cols}{the numerical column selected for a tour}
}
\value{
a list of edge linkage, edge color, and point color
}
\description{
A function to prepare edges data for tour display
}
