% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_cube}
\alias{profile_cube}
\title{Profile the data cube}
\usage{
profile_cube(dt)
}
\arguments{
\item{dt}{the data cube to profile}
}
\value{
here should be a customised proint of the profile
}
\description{
Profile the data cube
}
