% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-quality.R
\name{quality_plot}
\alias{quality_plot}
\title{data quality plot}
\usage{
quality_plot(data, ..., order_op = I)
}
\arguments{
\item{data}{the data set with missing calculated}

\item{...}{the variables to order by}

\item{order_op}{the operation on the variables for ordering, default to identity}
}
\description{
data quality plot
}
\examples{
# add missing summaries
out <- climate_small \%>\%
         add_missing_prct(prcp:tmin) \%>\%
         add_missing_dscrb()

# order by a single variable
out \%>\% quality_plot(prcp_missing)
out \%>\% quality_plot(tmax_missing)

# order by mean or median of several variables
out \%>\% quality_plot(prcp_missing, tmax_missing, order_op = mean)
out \%>\% quality_plot(prcp_missing, tmax_missing, tmin_missing, order_op = median)
}
