% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{rename_key}
\alias{rename_key}
\title{Rename the key variable}
\usage{
rename_key(data, ...)
}
\arguments{
\item{data}{a cubble}

\item{...}{argument passed to \code{rename}: NEW = OLD}
}
\description{
Rename the key variable
}
\examples{
aus_climate \%>\% rename_key(station = id)
}
