% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river}
\alias{river}
\title{Australia river data}
\format{
An object of class \code{cubble_df} (inherits from \code{rowwise_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 71 rows and 5 columns.
}
\usage{
river
}
\description{
Australia river data
}
\keyword{datasets}
