% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_group_var}
\alias{sample_group_var}
\title{Sample group variables for cubble}
\usage{
sample_group_var(data, size = 5, ...)
}
\arguments{
\item{data}{a cubble object to sample}

\item{size}{the number to sample, see \link[base]{sample}}

\item{...}{other argument passed to \link[base]{sample}}
}
\description{
Sample group variables for cubble
}
