% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{simplify_sf}
\alias{simplify_sf}
\title{Find (multi)polygons with small area}
\usage{
simplify_sf(data, geom, area, point_threshold = 0.9, area_threshold = 0.9)
}
\arguments{
\item{data}{An sf object}

\item{geom}{The geometry column}

\item{area}{The area column if any}

\item{point_threshold}{The number of point threshold used to define small crumb}

\item{area_threshold}{The area size threshold used to define small crumb}
}
\value{
An sf object
}
\description{
Find (multi)polygons with small area
}
