% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_head.cubble_df}
\alias{slice_head.cubble_df}
\alias{slice_tail.cubble_df}
\alias{slice_min.cubble_df}
\alias{slice_max.cubble_df}
\alias{slice_sample.cubble_df}
\title{Slicing a cubble}
\usage{
\method{slice_head}{cubble_df}(data, ...)

\method{slice_tail}{cubble_df}(data, ...)

\method{slice_min}{cubble_df}(data, ...)

\method{slice_max}{cubble_df}(data, ...)

\method{slice_sample}{cubble_df}(data, ...)
}
\arguments{
\item{data}{a cubble object to slice}

\item{...}{other arguments passed to the \code{\link[dplyr:slice]{dplyr::slice()}}}
}
\description{
Slicing a cubble
}
\examples{
\dontrun{
library(cubble)
# create a cubble object
all <- climate \%>\%
         dplyr::left_join(station, by = c("station" = "id")) \%>\%
         global(station)

# slice the first 50 stations from the top/ bottom
all \%>\% slice_head(n = 50)
all \%>\% slice_tail(n = 50)

# slice based on the max/ min of a variable
all \%>\% slice_max(elev, n = 10)
all \%>\% slice_min(lat, n = 10)

# random sample
all \%>\% slice_sample(n = 10)
}
}
