% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{slice_by}
\alias{slice_by}
\title{Slice the data cube into lower dimension}
\usage{
slice_by(dt, key)
}
\arguments{
\item{dt}{the raw dataframe to slice, in long form - does it has to be long form?}

\item{key}{the dimension to slice}
}
\value{
a nested tibble/ tsibble - think through here
}
\description{
Slice the data cube into lower dimension
}
