% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attrs.R
\name{spatial}
\alias{spatial}
\alias{spatial.spatial_cubble_df}
\alias{spatial.temporal_cubble_df}
\title{extract the spatial component from a cubble}
\usage{
spatial(data)

\method{spatial}{spatial_cubble_df}(data)

\method{spatial}{temporal_cubble_df}(data)
}
\arguments{
\item{data}{a cubble object}
}
\description{
extract the spatial component from a cubble
}
\examples{
spatial(climate_mel)
}
