% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{station}
\alias{station}
\title{Australian weather stations}
\format{
A tibble with 1,451 rows and 6 columns
\describe{
\item{id}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation station number, if has one}
}
}
\usage{
station
}
\description{
Australian weather stations
}
\seealso{
climate_large
}
\keyword{datasets}
