% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowwise.R
\name{switch_key}
\alias{switch_key}
\title{Switch to a different key of a cubble}
\usage{
switch_key(data, new_key, back = FALSE)
}
\arguments{
\item{data}{a cubble}

\item{new_key}{the new key}

\item{back}{whether it is a switch back to the individual level}
}
\description{
Switch to a different key of a cubble
}
\examples{
library(ggplot2)
# create a data with hierarchical structure:
# id (station) is nested in country, country is nested in continent
long <- world_climate \%>\%
  stretch() \%>\%
  migrate(country, continent) \%>\%
  filter(lubridate::year(date) == 2020)

# switch the key from id to country to aggregate on country
country_wise <- long \%>\%
  switch_key(country) \%>\%
  group_by(date) \%>\%
  summarise(tmax = mean(tmax, na.rm = TRUE))

# plot the aggregated maximum temperature for each country
country_wise \%>\%
  ggplot(aes(x = date, y = tmax, group = country)) +
  geom_line() +
  facet_wrap(vars(country)) +
  scale_x_date(date_labels = "\%b")
}
