% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-tamp.R
\name{tamp}
\alias{tamp}
\title{Switch an cubble object from the long form into the nested form}
\usage{
tamp(data)
}
\arguments{
\item{data}{a long cubble object}
}
\description{
\code{tamp()} turns a long cubble back into a nest cubble and can be seen as
the inverse operation of \code{stretch()}.
The nested cubble identifies each row by \code{key} and is suitable
for operations whose output doesn't involve a time index.
}
\examples{
cb_long <- climate_flat \%>\%
  as_cubble(key = id, index = date, coords = c(long, lat)) \%>\%
  stretch()

cb_long \%>\% tamp()
}
