% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{update_cubble}
\alias{update_cubble}
\alias{update_cubble.spatial_cubble_df}
\title{Temporary update cubble if the sf class take precedent of cubble classes}
\usage{
update_cubble(data, key, index, coords, ...)

\method{update_cubble}{spatial_cubble_df}(data, key = NULL, index = NULL, coords = NULL, ...)
}
\arguments{
\item{data, key, index, coords, ...}{see \code{make_cubble}}
}
\description{
When the data is already a cubble object but need update on attributes
}
