% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{world_climate}
\alias{world_climate}
\title{Worldwide climate data}
\format{
A tibble object with 177 rows and 9 columns
\describe{
\item{id}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{name of the station}
\item{wmo_id}{the World Meterological Organisation (WMO) station number}
\item{country}{country name}
\item{continent}{continent name}
\item{ts}{a list-column that nests all the time-wise measures: date, prcp, tmax, and tmin}
}
}
\usage{
world_climate
}
\description{
The dataset contains weather data from WMO and GSN stations in
the United States, Australia, Austria, Spain, France, and China.
}
\examples{
\dontrun{
library(ggplot2)
map <- sf::st_as_sf(rnaturalearth::ne_countries(scale = "medium", returnclass = "sf"))
plot_map(map) +
  geom_point(data = world_climate, aes(x = long, y = lat))
}

}
\keyword{datasets}
