% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Switch a cubble object into a long form}
\usage{
zoom(data, key)
}
\arguments{
\item{data}{a cubble object}

\item{key}{the spatio identifier. Key can be automatically detected for a cubble object}
}
\description{
Switch a cubble object into a long form
}
\details{
\code{zoom()} switch a cubble object into the long form where the combination of group identifier
and timestamp defines a row. The long form cubble is always of class \code{cubble_df} and \code{grouped_df}.
}
\examples{
climate_flat \%>\% global(station) \%>\% zoom()

}
\seealso{
Other cubble verbs include \code{\link{global}} and \code{\link{migrate}}
}
