% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{add_geometry_column}
\alias{add_geometry_column}
\title{add geometry list column to cubble_df object}
\usage{
add_geometry_column(x, sfc = NULL, crs, silent = FALSE)
}
\arguments{
\item{x}{object of class \code{cubble}}

\item{sfc}{object of class \code{sfc} (see package sf)}

\item{crs}{object of class \code{crs} (see package sf); if missing 'OGC:CRS84' is assumed (WGS84) and a message is emitted}

\item{silent}{logical; suppress message?}
}
\description{
add geometry list column to cubble_df object
}
\examples{
climate_mel \%>\% add_geometry_column()
}
