% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_time}
\alias{aggregate_time}
\title{Aggregate the time axis}
\usage{
aggregate_time(dt, val, unit, op)
}
\arguments{
\item{dt}{the data cube}

\item{val}{the value column to aggregate}

\item{unit}{the time unit to aggregate into, currently support "y", "ym", "ymd"}

\item{op}{a function to summarise by}
}
\value{
a cube with index being aggregated to the instructed value
}
\description{
Aggregate the time axis
}
\examples{
\dontrun{
 water <- as_cube(water_raw, key = station, index = time, var = parameter)
 a <- aggregate_time(water, value, "ymd", mean)
}
}
