% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{as_cube}
\alias{as_cube}
\title{casting and coercion}
\usage{
as_cube(x, key = NULL, index, var = NULL, ...)
}
\description{
casting and coercion
}
\examples{
\dontrun{
water <- as_cube(water_small, index = date, key = station, var = parameter)
attributes(water)
ped_cube <- as_cube(pedestrian, var = Count)
attributes(ped_cube)
}
}
