% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-attr.R
\name{form}
\alias{form}
\alias{leaves}
\alias{group_vars}
\alias{groups}
\title{Functions to get attributes from a cubble object}
\usage{
form(data)

leaves(data)

group_vars(data)

groups(data)
}
\arguments{
\item{data}{an cubble object}
}
\description{
Functions to get attributes from a cubble object
}
\details{
Apart from inheriting attributes \code{names}, \code{row.names}, and \code{class} from the underlying tibble,
a cubble object will also have additional attributes: \code{groups}, \code{meta} and \code{form}.

\code{groups()}, \code{meta()}, and \code{form()} provide a more  convenient way to extract
these attributes without \verb{\%@\%} from a cubble.
\code{group_vars()} extracts the variable name of the group, which can often be useful.

If a cubble object is also a tsibble, then tsibble attributes (\code{key}, \code{index}, \code{index2}, \code{interval})
are also preserved and can be accessed via the relevant functions in the tsibble package.
}
\examples{
# extract attributes of a cubble object
form(climate_small)
leaves(climate_small) \%>\% head(5)
groups(climate_small) \%>\% head(5)
group_vars(climate_small)

# print out the attribute names of cubble with a tsibble underlying class
names(attributes(climate_small \%>\% stretch()))
}
