% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox.R
\name{bbox}
\alias{bbox}
\title{Extract bbox of the coordinate with a buffer}
\usage{
bbox(data, lat, long, buffer_ratio = 0.01)
}
\arguments{
\item{data}{the data object}

\item{lat}{the latitude column}

\item{long}{the longitude column}

\item{buffer_ratio}{the percentage to add/deduct to the max/min lat/long when querying the map}
}
\description{
Extract bbox of the coordinate with a buffer
}
\examples{
library(ggmap)
library(ggplot2)
aus_map <- get_map(location = bbox(station), source = "osm")
ggmap(aus_map) + geom_point(data = station, aes(x = long, y = lat))
}
