% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_flat}
\alias{climate_flat}
\title{Australia climate data (minimal)}
\format{
A tibble object with 732 rows and 9 columns
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{prcp}{precipitation}
\item{tmax}{maximum temperature}
\item{tmin}{minimum temperature}
}
}
\usage{
climate_flat
}
\description{
A minimal 2020 climate dataset recorded in 2 stations in a tibble format. This data is used to
show how to create a cubble object from tibble.
}
\seealso{
climate_small
}
\keyword{datasets}
