% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_flat}
\alias{climate_flat}
\title{Australia climate data (small)}
\format{
A tibble object with 120, 321 rows and 9 columns
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{prcp}{precipitation}
\item{tmax}{maximum temperature}
\item{tmin}{minimum temperature}
}
}
\usage{
climate_flat
}
\description{
Climate data recorded in 55 stations from 2015 - 2020 in Australia.
This data has the same content as \link{climate_small} but in a tibble format.
}
\seealso{
climate_small
}
\keyword{datasets}
