% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_subset}
\alias{climate_subset}
\title{Australia climate data - 30 stations}
\format{
A cubble object
\describe{
\item{id}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{name}{station name}
\item{wmo_id}{the world meterological organisation (WMO) station number}
\item{ts}{a list-column that nests all the time-wise measures: date, prcp, tmax, and tmin}
}
}
\usage{
climate_subset
}
\description{
Daily measure on precipitation (prcp) maximum temperature (tmax), and
minimum temperature (tmin) in 2020 for 30 stations.
}
\examples{
library(ggplot2)
state_map <- rmapshaper::ms_simplify(ozmaps::abs_ste, keep = 2e-3)
ggplot2::ggplot() +
  ggplot2::geom_sf(data = state_map,
                   ggplot2::aes(geometry = .data$geometry),
                   color = "grey", linetype = "dotted") +
  ggplot2::geom_point(data = climate_subset,
                      ggplot2::aes(x = long, y = lat)) +
  ggplot2::theme_bw()
}
\seealso{
climate_aus climate_flat
}
\keyword{datasets}
