% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R, R/cubble-s3.R
\name{is_cubble}
\alias{is_cubble}
\alias{cubble}
\title{Create a cubble object}
\usage{
is_cubble(data)

cubble(..., key, index, coords)
}
\arguments{
\item{data}{object to be converted into an object of class \code{cubble_df}}

\item{...}{a list object to create new cubble}

\item{key}{the variable(s) that identifies the spatial location.}

\item{index}{the single variable that identifies time, currently support
base R classes \code{Date}, \code{POSIXlt}, \code{POSIXct} and
tsibble's \code{yearmonth}, \code{yearweek}, and \code{yearquarter} class,
supply in non-standard evaluation}

\item{coords}{the coordinate columns, in the form of \code{c(LONGITUDE, LATITUDE)}
the argument can be omitted if created from an sf and its subclasses.
In case the sf geometry column is not POINT, cubble will use the centroid
coordinates as LONGITUDE and LATITUDE}
}
\value{
a TRUE/FALSE predicate

a cubble object
}
\description{
Create a cubble object
}
