% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cubble.R, R/cubbe-s3.R
\name{as_cubble}
\alias{as_cubble}
\alias{as_cubble.tbl_df}
\alias{as_cubble.rowwise_df}
\alias{cubble}
\alias{tbl_sum.cubble_df}
\alias{is_cubble}
\title{Constructor for the cubble class}
\usage{
as_cubble(data, key, index, coords)

\method{as_cubble}{tbl_df}(data, key, index, coords)

\method{as_cubble}{rowwise_df}(data, key, index, coords)

cubble(..., key, index, coords)

\method{tbl_sum}{cubble_df}(data)

is_cubble(data)
}
\arguments{
\item{data}{the object to be created or tested as cubble}

\item{key}{the spatial identifier}

\item{index}{the time identifier}

\item{coords}{the coordinates that characterise the spatial dimension}

\item{...}{a list object to create new cubble}
}
\description{
Constructor for the cubble class
}
\examples{
climate_flat \%>\%
  as_cubble(key = id, index = date, coords = c(long, lat))
}
