% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cubble.R, R/cubbe-s3.R
\name{as_cubble}
\alias{as_cubble}
\alias{as_cubble.tbl_df}
\alias{as_cubble.rowwise_df}
\alias{cubble}
\alias{tbl_sum.cubble_df}
\alias{is_cubble}
\title{Constructor for the cubble class}
\usage{
as_cubble(data, key, index, coords)

\method{as_cubble}{tbl_df}(data, key, index, coords)

\method{as_cubble}{rowwise_df}(data, key, index, coords)

cubble(..., key, index, coords)

\method{tbl_sum}{cubble_df}(data)

is_cubble(data)
}
\arguments{
\item{data}{the object to be created or tested as cubble}

\item{key}{the spatial identifier}

\item{index}{the time identifier}

\item{coords}{the coordinates that characterise the spatial dimension}

\item{...}{a list object to create new cubble}
}
\description{
Constructor for the cubble class
}
\examples{
# create a nested tibble
library(tibble)
dt <- tibble(
  station = c("A", "B"),
  long = c(110, 120),
  lat = c(-10, -20),
  ts = list(
    tibble(
      date = c(
        as.Date("2021-07-01"),
        as.Date("2021-08-01")
      ),
      prcp = c(0, 0),
      tmax = c(20, 125),
      tmin = c(15, 18)
    ),
    tibble(
      date = c(
        as.Date("2021-07-01"),
        as.Date("2021-08-01")
      ),
      prcp = c(0, 100),
      tmax = c(10, 15),
      tmin = c(5, 8)
    )
  )
)

# create a cubble from `dt`
# cubble(dt,
#        group = station,
#        leaves = tibble(
#          station = c("A", "B"),
#          long = c(110, 120),
#          lat = c(-10, -20)
#        ),
#        form = "long"
# )
}
