% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubbe-s3.R
\name{cubble}
\alias{cubble}
\alias{new_cubble}
\alias{tbl_sum.cubble_df}
\alias{is_cubble}
\title{Constructor for the cubble class}
\usage{
cubble(..., group, leaves, form)

new_cubble(data, group, leaves, form)

\method{tbl_sum}{cubble_df}(data)

is_cubble(data)
}
\arguments{
\item{...}{a list object to create new cubble}

\item{group}{the spatio identifier}

\item{leaves}{metadata to include in the attributes}

\item{form}{whether the long or wide form}

\item{data}{the object to be created or tested as cubble}
}
\description{
Constructor for the cubble class
}
\examples{
# create a nested tibble
dt <- tibble(
  station = c("A", "B"),
  long = c(110, 120),
  lat = c(-10, -20),
  ts = list(
    tibble(
      date = c(
        as.Date("2021-07-01"),
        as.Date("2021-08-01")
      ),
      prcp = c(0, 0),
      tmax = c(20, 125),
      tmin = c(15, 18)
    ),
    tibble(
      date = c(
        as.Date("2021-07-01"),
        as.Date("2021-08-01")
      ),
      prcp = c(0, 100),
      tmax = c(10, 15),
      tmin = c(5, 8)
    )
  )
)

# create a cubble from `dt`
cubble(dt,
       group = station,
       leaves = tibble(
         station = c("A", "B"),
         long = c(110, 120),
         lat = c(-10, -20)
       ),
       form = "long"
)
}
