% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-class.R
\name{cubble}
\alias{cubble}
\alias{make_cubble}
\title{Create a cubble object}
\usage{
cubble(..., key, index, coords)

make_cubble(spatial, temporal, by = NULL, key, index, coords)
}
\arguments{
\item{...}{a list object to create new cubble}

\item{key}{the variable(s) that identifies the spatial location.}

\item{index}{the single variable that identifies time, currently support
base R classes \code{Date}, \code{POSIXlt}, \code{POSIXct} and
tsibble's \code{yearmonth}, \code{yearweek}, and \code{yearquarter} class,
supply in non-standard evaluation}

\item{coords}{the coordinate columns, in the form of \code{c(LONGITUDE, LATITUDE)}
the argument can be omitted if created from an sf and its subclasses.
In case the sf geometry column is not POINT, cubble will use the centroid
coordinates as LONGITUDE and LATITUDE}

\item{spatial}{a tibble of spatial variables, contain key and the two coords columns}

\item{temporal}{a tibble of temporal variables, contain key and index}

\item{by}{used in \code{make_cubble} when the key variable has different names in the
spatial and temporal data, in the syntax of the \code{by} argument in \code{left_join}  (see examples)}
}
\value{
a cubble object
}
\description{
Create a cubble object
}
\examples{
cubble(
  id = rep(c("perth", "melbourne", "sydney"), each = 3),
  date = rep(as.Date("2020-01-01") + 0:2, times = 3),
  long = rep(c(115.86, 144.96, 151.21), each = 3),
  lat = rep(c(-31.95, -37.81, -33.87), each = 3),
  value = rnorm(n = 9),
  key = id, index = date, coords = c(long, lat)
  )

# stations and climate are in-built data in cubble
make_cubble(spatial = stations, temporal = meteo,
            key = id, index = date, coords = c(long, lat))

}
