% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-cubble.R, R/cubble-package.R
\docType{package}
\name{cubble}
\alias{cubble}
\alias{new_cubble}
\alias{as_cubble}
\alias{cubble-package}
\alias{_PACKAGE}
\title{Create a cubble object}
\usage{
cubble(main, item, by = NULL)

new_cubble(main, item, by)

as_cubble()
}
\arguments{
\item{main}{the main sheet (a tsibble)}

\item{item}{the item sheet (a tibble)}

\item{by}{the variable that join the two sheets}
}
\description{
A spatiotemperal data object in a relational data structure to separate the recording of time variant/ invariant variables.
}
\examples{
df <- cubble(climate, station, by = c("station" = "id"))
}
\author{
\strong{Maintainer}: Sherry Zhang \email{huize.zhang@monash.edu} (\href{https://orcid.org/0000-0002-7122-1463}{ORCID})

}
\keyword{internal}
