% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubble-print.R
\name{print.cubble_df}
\alias{print.cubble_df}
\alias{tbl_sum.cubble_df}
\alias{tbl_sum.spatial_cubble_df}
\alias{tbl_sum.temporal_cubble_df}
\title{print a cubble object}
\usage{
\method{print}{cubble_df}(
  x,
  width = NULL,
  ...,
  n_extra = NULL,
  n = NULL,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

\method{tbl_sum}{cubble_df}(x)

\method{tbl_sum}{spatial_cubble_df}(x)

\method{tbl_sum}{temporal_cubble_df}(x)
}
\arguments{
\item{x, width, n_extra, n, max_extra_cols, max_footer_lines}{see pillar tbl-format.R}

\item{...}{other argument to pass into \code{format()}}
}
\description{
print a cubble object
}
