% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrate.R
\name{migrate}
\alias{migrate}
\title{Move item-wise variables into long form}
\usage{
migrate(data, ...)
}
\arguments{
\item{data}{a long form tibble object}

\item{...}{variables in metadata to included into the long form}
}
\description{
Move item-wise variables into long form
}
\details{
Some spatio-temporal visualisations, i.e. a glyph map, require both item-wise
and time-wise variables to create the plot. \code{migrate()} moves item-wise variables
into the long form before creating the visualisation.
}
\examples{
library(ggplot2)
library(GGally)
library(dplyr)
dt <- aus_climate \%>\%
  stretch() \%>\%
  migrate(lat, long) \%>\%
  filter(lubridate::month(date) == 2 ) \%>\%
  mutate(yday = lubridate::yday(date)) \%>\%
  select(-date)

g <- glyphs(dt,
            x_major = "long", x_minor = "yday",
            y_major = "lat", y_minor = "tmax", height = 1, width = 2)

state_map <- rmapshaper::ms_simplify(ozmaps::abs_ste, keep = 2e-3)
plot_map(state_map) +
  ggplot2::geom_path(data = g, aes(gx, gy, group = gid)) +
  ggplot2::theme_void() +
  ggplot2::labs(x = "", y = "")
}
