% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-zoom.R
\name{global}
\alias{global}
\alias{zoom}
\title{Data structure}
\usage{
global(data, key)

zoom(data, col)
}
\arguments{
\item{data}{the data to zoom}

\item{key}{the spatio identifier}

\item{col}{the list-column in the data to zoom into}
}
\description{
Data structure
}
\examples{
oz_global <- global(oz_climate, station)
oz_zoom <- oz_global \%>\% zoom(data)
back <- oz_zoom \%>\% global(station)
}
