% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-zoom.R
\name{global}
\alias{global}
\alias{zoom}
\title{Data structure}
\usage{
global(data, key)

zoom(data, col)
}
\description{
Data structure
}
\examples{
global <- global(oz_climate, station)
zoom <- global \%>\% zoom(data)
}
