% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-zoom.R
\name{global}
\alias{global}
\alias{cubble_df}
\alias{new_cubble_df}
\alias{zoom}
\alias{is_cubble}
\title{Data structure}
\usage{
global(data, key)

cubble_df(data, group, meta_data, form)

new_cubble_df(data, group, meta_data, form)

zoom(data)

is_cubble(data)
}
\arguments{
\item{data}{the data to be converted into a cubble object}

\item{key}{the spatio identifier}

\item{group}{the spatio identifier}

\item{meta_data}{metadata to include in the attributes}

\item{form}{whether the long or wide form}

\item{col}{the list-column in the data to zoom into}
}
\description{
Data structure
}
\examples{
oz_global <- global(oz_climate, station)
oz_zoom <- oz_global \%>\% zoom()
back <- oz_zoom \%>\% global(station)
}
