% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubbe-s3.R
\name{cubble}
\alias{cubble}
\alias{new_cubble}
\alias{is_cubble}
\title{Extract cubble attributes}
\usage{
cubble(..., group, leaves, form)

new_cubble(data, group, leaves, form)

is_cubble(data)
}
\arguments{
\item{...}{a list object to create new cubble}

\item{group}{the spatio identifier}

\item{leaves}{metadata to include in the attributes}

\item{form}{whether the long or wide form}

\item{data}{the object to be created or tested as cubble}
}
\description{
Extract cubble attributes
}
