% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-zoom.R
\name{global}
\alias{global}
\alias{cubble_df}
\alias{new_cubble_df}
\alias{zoom}
\alias{is_cubble}
\title{Data structure}
\usage{
global(data, key)

cubble_df(data, group, meta_data, form)

new_cubble_df(data, group, meta_data, form)

zoom(data, key)

is_cubble(data)
}
\arguments{
\item{data}{the data to be converted into a cubble object}

\item{key}{the spatio identifier. key can be automatically detected for a cubble object}

\item{group}{the spatio identifier}

\item{meta_data}{metadata to include in the attributes}

\item{form}{whether the long or wide form}
}
\description{
Data structure
}
\examples{
\dontrun{
oz_global <- global(oz_climate, station)
oz_zoom <- oz_global \%>\% zoom()
back <- oz_zoom \%>\% global()


oz_climate_tsibble <- oz_climate \%>\% as_tsibble(index = date, key = station)
oz_global2 <- oz_climate_tsibble \%>\% global(station)
oz_zoom2 <- oz_global2 \%>\% zoom()
back <- oz_zoom2 \%>\% global()

outer <- water_small \%>\%
 global(station) \%>\%
 mutate(river_name = stringr::str_extract(name, "(^.+)(?=( RIVER| CREEK))"),
   river_id = stringr::str_sub(river_name, 1, 1)) \%>\%
 global(key = river_name)

mid <- outer \%>\% zoom()
inner <- mid \%>\% zoom()

# now convert inner back will cause station-wise variables: station, name, lat, and long to lost
back <- inner \%>\% global()
meta(back)
}

}
