% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cubbe-s3.R
\name{cubble_df}
\alias{cubble_df}
\alias{new_cubble_df}
\alias{is_cubble}
\title{Extract cubble attributes}
\usage{
cubble_df(data, group, meta_data, form)

new_cubble_df(data, group, meta_data, form)

is_cubble(data)
}
\arguments{
\item{data}{the object to be created or tested as cubble}

\item{group}{the spatio identifier}

\item{meta_data}{metadata to include in the attributes}

\item{form}{whether the long or wide form}
}
\description{
Extract cubble attributes
}
