% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{arrange.temporal_cubble_df}
\alias{arrange.temporal_cubble_df}
\alias{select.spatial_cubble_df}
\alias{select.temporal_cubble_df}
\alias{group_by.spatial_cubble_df}
\alias{group_by.temporal_cubble_df}
\alias{summarise.spatial_cubble_df}
\alias{summarise.temporal_cubble_df}
\alias{rename.spatial_cubble_df}
\alias{rename.temporal_cubble_df}
\alias{bind_rows.temporal_cubble_df}
\alias{bind_cols.spatial_cubble_df}
\alias{bind_cols.temporal_cubble_df}
\alias{rowwise.spatial_cubble_df}
\alias{rowwise.temporal_cubble_df}
\alias{dplyr_col_modify.cubble_df}
\alias{dplyr_row_slice.spatial_cubble_df}
\alias{dplyr_row_slice.temporal_cubble_df}
\alias{dplyr_reconstruct.spatial_cubble_df}
\alias{dplyr_reconstruct.temporal_cubble_df}
\alias{mutate.spatial_cubble_df}
\alias{filter.spatial_cubble_df}
\alias{arrange.spatial_cubble_df}
\title{Access to dplyr verbs}
\usage{
\method{arrange}{temporal_cubble_df}(.data, ...)

\method{select}{spatial_cubble_df}(.data, ...)

\method{select}{temporal_cubble_df}(.data, ...)

\method{group_by}{spatial_cubble_df}(.data, ..., .add, .drop)

\method{group_by}{temporal_cubble_df}(.data, ..., .add, .drop)

\method{summarise}{spatial_cubble_df}(.data, ..., .by = NULL, .groups = NULL)

\method{summarise}{temporal_cubble_df}(.data, ..., .by = NULL, .groups = NULL)

\method{rename}{spatial_cubble_df}(.data, ...)

\method{rename}{temporal_cubble_df}(.data, ...)

bind_rows.temporal_cubble_df(..., .id = NULL)

bind_cols.spatial_cubble_df(..., .name_repair)

bind_cols.temporal_cubble_df(..., .name_repair)

\method{rowwise}{spatial_cubble_df}(data, ...)

\method{rowwise}{temporal_cubble_df}(data, ...)

\method{dplyr_col_modify}{cubble_df}(data, cols)

\method{dplyr_row_slice}{spatial_cubble_df}(data, i, ...)

\method{dplyr_row_slice}{temporal_cubble_df}(data, i, ...)

\method{dplyr_reconstruct}{spatial_cubble_df}(data, template)

\method{dplyr_reconstruct}{temporal_cubble_df}(data, template)

\method{mutate}{spatial_cubble_df}(.data, ...)

\method{filter}{spatial_cubble_df}(.data, ...)

\method{arrange}{spatial_cubble_df}(.data, ...)
}
\arguments{
\item{data, .data}{a cubble object(used as defined by the dplyr generic)}

\item{cols, i, template, ..., .add, .drop, .id, .name_repair, .by, .groups}{see docs in dplyr verbs}
}
\description{
Verbs supported for both nested and long cubble include:
\code{\link[dplyr:mutate]{dplyr::mutate()}}, \code{\link[dplyr:filter]{dplyr::filter()}},  \code{\link[dplyr:arrange]{dplyr::arrange()}}, \code{\link[dplyr:select]{dplyr::select()}},
\code{\link[dplyr:group_by]{dplyr::group_by()}}, \code{\link[dplyr:summarise]{dplyr::summarise()}},. \code{\link[dplyr:rename]{dplyr::rename()}}, \code{\link[dplyr:bind_cols]{dplyr::bind_cols()}},
\code{dplyr::slice_*()}, \code{dplyr::*_join()}, \code{\link[dplyr:relocate]{dplyr::relocate()}}, \code{\link[dplyr:pull]{dplyr::pull()}}
}
\examples{
#today
}
\references{
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
