% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\name{arrange.temporal_cubble_df}
\alias{arrange.temporal_cubble_df}
\alias{select.spatial_cubble_df}
\alias{select.temporal_cubble_df}
\alias{group_by.spatial_cubble_df}
\alias{group_by.temporal_cubble_df}
\alias{summarise.spatial_cubble_df}
\alias{summarise.temporal_cubble_df}
\alias{rename.spatial_cubble_df}
\alias{rename.temporal_cubble_df}
\alias{bind_rows.temporal_cubble_df}
\alias{bind_cols.spatial_cubble_df}
\alias{bind_cols.temporal_cubble_df}
\alias{dplyr_col_modify.cubble_df}
\alias{dplyr_row_slice.spatial_cubble_df}
\alias{dplyr_row_slice.temporal_cubble_df}
\alias{dplyr_reconstruct.spatial_cubble_df}
\alias{dplyr_reconstruct.temporal_cubble_df}
\alias{mutate.spatial_cubble_df}
\alias{filter.spatial_cubble_df}
\alias{arrange.spatial_cubble_df}
\title{Access to dplyr verbs}
\usage{
\method{arrange}{temporal_cubble_df}(.data, ...)

\method{select}{spatial_cubble_df}(.data, ...)

\method{select}{temporal_cubble_df}(.data, ...)

\method{group_by}{spatial_cubble_df}(.data, ..., .add, .drop)

\method{group_by}{temporal_cubble_df}(.data, ..., .add, .drop)

\method{summarise}{spatial_cubble_df}(.data, ..., .by = NULL, .groups = NULL)

\method{summarise}{temporal_cubble_df}(.data, ..., .by = NULL, .groups = NULL)

\method{rename}{spatial_cubble_df}(.data, ...)

\method{rename}{temporal_cubble_df}(.data, ...)

bind_rows.temporal_cubble_df(..., .id = NULL)

bind_cols.spatial_cubble_df(..., .name_repair)

bind_cols.temporal_cubble_df(..., .name_repair)

\method{dplyr_col_modify}{cubble_df}(data, cols)

\method{dplyr_row_slice}{spatial_cubble_df}(data, i, ...)

\method{dplyr_row_slice}{temporal_cubble_df}(data, i, ...)

\method{dplyr_reconstruct}{spatial_cubble_df}(data, template)

\method{dplyr_reconstruct}{temporal_cubble_df}(data, template)

\method{mutate}{spatial_cubble_df}(.data, ...)

\method{filter}{spatial_cubble_df}(.data, ...)

\method{arrange}{spatial_cubble_df}(.data, ...)
}
\arguments{
\item{.by, .groups}{used by dplyr verbs}

\item{data, .data}{a cubble object(used as defined by the dplyr generic)}

\item{cols, i, template, ..., .add, .drop, .id, .name_repair}{see \code{\link[dplyr:dplyr_extending]{dplyr::dplyr_col_modify()}}, \code{\link[dplyr:dplyr_extending]{dplyr::dplyr_row_slice()}},
and \code{\link[dplyr:dplyr_extending]{dplyr::dplyr_reconstruct()}}}
}
\description{
Access to dplyr verbs
}
\references{
https://dplyr.tidyverse.org/reference/dplyr_extending.html
}
