% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-face_temporal.R
\name{face_temporal}
\alias{face_temporal}
\title{Switch a cubble object from the nested form into the long form}
\usage{
face_temporal(data, col)
}
\arguments{
\item{data}{a nested cubble object}

\item{col}{the list column to be face_temporaled, \code{col} is required to be specified
if there are more than one list column and the list column name is not \code{ts}}
}
\description{
\code{face_temporal()} switches a cubble object into a long cubble, suitable for temporal operations.
The long cubble uses the combination of \code{key} and \code{index} to identify each row and
arranges each \code{key} as a separate group.
}
\examples{
climate_flat \%>\%
  as_cubble(key = id, index = date, coords = c(long, lat)) \%>\%
  face_temporal()
}
