% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{gen_grid}
\alias{gen_grid}
\title{Automatic generation of geofacet grid}
\usage{
gen_grid(data, var = NULL, nrow = NULL, ncol = NULL)
}
\arguments{
\item{data}{a cubble object}

\item{var}{the facet variable}

\item{nrow}{number of row in the facet}

\item{ncol}{number of column in the facet}
}
\description{
This function automatically generates a grid for \code{\link[geofacet]{facet_geo}}
based on longitude and latitude.
}
\examples{
library(lubridate)
library(ggplot2)
library(geofacet)

# filter out 15 stations in Victoria
small <- victoria \%>\% slice_tail(n = 15)

####################
# how those stations look like in Victoria map
state_map <- ms_simplify(ozmaps::abs_ste \%>\%
                           filter(NAME == "Victoria"),
                         keep = 2e-3)
plot_map(state_map) +
  geom_point(data = short, aes(x = long, y = lat)) +
  ggrepel::geom_label_repel(data = short, aes(x = long, y = lat, label = name))

####################
# generate the grid
grid1 <- short \%>\% gen_grid(var = name)

# looks like there's a sudden tmax drop around Feb 2020 around melbourne
# (melbourne airport, wallan, essendon, avalon, and moorabbin airport) while
# the drop is less obvious in eastern victoria.
short \%>\%
  stretch() \%>\%
  migrate(name) \%>\%
  dplyr::filter(year(date) == 2020, month(date) == 2) \%>\%
  ggplot(aes(x = date, y = tmax, group = name)) +
  geom_line() +
  facet_geo(vars(name), grid = grid1)
}
