% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaves.R
\name{new_leaves}
\alias{new_leaves}
\alias{is_leaves}
\alias{invariant}
\alias{variant}
\title{Create a leaf object}
\usage{
new_leaves(data, group)

is_leaves(data)

invariant(data)

variant(data)
}
\arguments{
\item{data}{the object to be created or tested as cubble}

\item{group}{the spatio identifier}
}
\description{
Create a leaf object
}
