% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{make_temporal_tsibble}
\alias{make_temporal_tsibble}
\title{Update the temporal cubble to include the tsibble class (\code{tbl_ts})}
\usage{
make_temporal_tsibble(x)
}
\arguments{
\item{x}{object of class \code{temporal_cubble_df}}
}
\description{
Update the temporal cubble to include the tsibble class (\code{tbl_ts})
}
\examples{
climate_mel \%>\% face_temporal() \%>\% make_temporal_tsibble()
}
