% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{add_missing_prct}
\alias{add_missing_prct}
\title{Compute missing summary}
\usage{
add_missing_prct(data, vars)
}
\arguments{
\item{data}{a cubble object}

\item{vars}{variables to compute percentage missing (support tidyselect)}
}
\value{
a cubble object with additional columns VAR_missing
}
\description{
Compute missing summary
}
\details{
\itemize{
\item \code{add_missing_prct()} computes the percentage of missing for the selected variables
}
}
\examples{
climate_aus \%>\%  add_missing_prct(vars = prcp)
}
