% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{add_missing_prct}
\alias{add_missing_prct}
\alias{add_missing_dscrb}
\title{Compute statistics for variable missingness}
\usage{
add_missing_prct(data, ...)

add_missing_dscrb(data, cutoff = 0.99)
}
\arguments{
\item{data}{a cubble object}

\item{...}{the variable to compute \code{add_missing_prct()}}

\item{cutoff}{the threshold value, above which the variable
will be described as "almost all missing". Only used in \code{add_missing_dscrb()}}
}
\description{
Compute statistics for variable missingness
}
\examples{

\dontrun{
out <- oz_global2 \%>\% add_missing_prct(vars = c("prcp", "tmin", "tmax"))
out2 <- out \%>\% add_missing_dscrb()
}
}
