% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{add_missing_prct}
\alias{add_missing_prct}
\alias{add_missing_dscrb}
\title{Compute missing summary}
\usage{
add_missing_prct(data, ...)

add_missing_dscrb(data, cutoff = 0.99)
}
\arguments{
\item{data}{a cubble object}

\item{...}{variables to compute percentage missing (support tidyselect)}

\item{cutoff}{the threshold value, above which the variable
will be described as "almost all missing".}
}
\description{
Compute missing summary
}
\details{
\itemize{
\item \code{add_missing_prct()} computes the percentage of missing for the selected variables
\item \code{add_missing_dscrb()} gives a descriptive label (no missing, partly missing, and almost all missing)
for those variables that have missing percentage computed
}
}
\examples{
climate_small \%>\% add_missing_prct(prcp:tmin)
climate_small \%>\% add_missing_prct(prcp:tmin) \%>\% add_missing_dscrb()
}
