% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netcdf.R
\name{extract_var}
\alias{extract_var}
\alias{extract_longlat}
\alias{extract_time}
\title{Functions to extract NetCDF dimension and variables}
\usage{
extract_var(data, vars)

extract_longlat(data)

extract_time(data)
}
\arguments{
\item{data}{a NetCDF file read in from \code{ncdf4::nc_open()}}

\item{vars}{variables to read, see the variables in your data with \code{names(data$var)}}
}
\description{
Functions to extract NetCDF dimension and variables
}
