% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single-cb.R
\name{new_sheet_single}
\alias{new_sheet_single}
\title{Create a single cubble object}
\usage{
new_sheet_single(single_sheet, by, cb)
}
\arguments{
\item{single_sheet}{one of the sheet in a cubble object}

\item{by}{the joining variable}

\item{cb}{the other sheet in the cubble object}
}
\description{
Create a single cubble object
}
