% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oz_climate}
\alias{oz_climate}
\title{Australia climate data (dirty)}
\format{
A tibble object with 2,069,634 rows and 9 columns
\describe{
\item{station}{station id}
\item{lat}{latitude of the station}
\item{long}{longitude of the station}
\item{elev}{elevation of the station}
\item{prcp}{precipitation}
\item{tmax}{maximum temperature}
\item{tmin}{minimum temperature}
}
}
\usage{
oz_climate
}
\description{
Australia climate data (dirty)
}
\seealso{
climate_small
}
\keyword{datasets}
