% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{plot_map}
\alias{plot_map}
\title{A quick plot of maps}
\usage{
plot_map(map_data, ...)
}
\arguments{
\item{map_data}{the dataset contains the map to be plotted}

\item{...}{other arguments to be passed into \code{geom_sf()}}
}
\description{
A quick initiate of a ggplot2 object for map without typing
\code{data \%>\% ggplot2::ggplot() + ggplot2::geom_sf(ggplot2::aes(geometry = geometry))}
everytime to create a map.
}
\examples{
library(rmapshaper)
state_map <- ms_simplify(ozmaps::abs_ste, keep = 2e-3)
plot_map(state_map)
}
