% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-auto.R
\name{plot_map}
\alias{plot_map}
\title{A quick plot of map}
\usage{
plot_map(map_data, ...)
}
\arguments{
\item{map_data}{the dataset contains the map to be plotted}

\item{...}{other arguments to be passed into \code{geom_sf()}}
}
\description{
A quick initiate of a ggplot2 object for map with some default aesthetic setting
to save some typing for create a map.
}
\examples{
library(rmapshaper)
library(ggplot2)
library(dplyr)

vic_map <- ozmaps::abs_ste \%>\% filter(NAME == "Victoria")
plot_map(vic_map)

# Victoria stations
victoria <- aus_climate \%>\% filter(stringr::str_detect(id, "ASN0008"))
plot_map(vic_map) +
   geom_point(data = victoria, aes(x = long, y = lat))
}
