% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-auto.R
\name{plot_map}
\alias{plot_map}
\title{A quick plot of map}
\usage{
plot_map(map_data, point_data, print_code = FALSE)
}
\arguments{
\item{map_data}{the dataset contains the map object, an sf object}

\item{point_data}{a cubble object to plot the site}

\item{print_code}{whether to print out the ggplot2 code}
}
\description{
A quick plotting function with some default aesthetic and theme settings for
maps. To add more customisation on the plot, set \code{print_code = TRUE},
and modify in the relevant places.
}
\examples{
library(ggplot2)
state_map <- rmapshaper::ms_simplify(ozmaps::abs_ste, keep = 2e-3)
# a quick plot
plot_map(state_map, climate_aus)

# print out the ggplot2 code of the map
plot_map(state_map, climate_aus, print_code = TRUE)
}
