% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-slice.R
\name{slice_nearby}
\alias{slice_nearby}
\title{Location-based slicing}
\usage{
slice_nearby(data, coords, buffer, n)
}
\arguments{
\item{buffer}{the buffer added to the coordinate for slicing}

\item{n}{the number of nearby points to slice, based on distance}

\item{coord}{the coordinate of used to slice nearby locations}
}
\description{
Location-based slicing
}
\examples{
# slice locations within 1 degree of (130E, 25S)
slice_nearby(climate_aus, coord = c(130, -25), buffer = 3)

# slice the 5 closest location to (130E, 25S)
slice_nearby(climate_aus, coord = c(130, -25), n = 5)
}
