% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch.R
\name{stretch}
\alias{stretch}
\title{Switch a cubble object into a long form}
\usage{
stretch(data, key)
}
\arguments{
\item{data}{a cubble object}

\item{key}{the spatio identifier. Key can be automatically detected for a cubble object}
}
\description{
Switch a cubble object into a long form
}
\details{
\code{stretch()} switch a cubble object into the long form where the combination of group identifier
and timestamp defines a row. The long form cubble is always of class \code{cubble_df} and \code{grouped_df}.
}
\examples{
climate_flat \%>\% tamp(station) \%>\% stretch()

}
\seealso{
Other cubble verbs include \code{\link{tamp}} and \code{\link{migrate}}
}
