% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tamp.R
\name{tamp}
\alias{tamp}
\title{Formulate or switch an object into a cubble in the nested form}
\usage{
tamp(data, key)
}
\arguments{
\item{data}{the data to be converted into a cubble object}

\item{key}{the spatio identifier. Key can be automatically detected for a cubble object}
}
\description{
Create a cubble object or switch an existing cubble object into the nested form
}
\details{
\itemize{
\item For an object that is not \code{cubble_df}, \code{tamp()} will initialise a cubble object upon
providing a \code{group} variable.
\item For an existing \code{cubble} in the long form, \code{tamp()}
switches the object back to the nested form. This form is easier to computing
group-specific variables.
}
}
\examples{
# switch to the nested form from the long form
aus_climate \%>\%
  stretch() \%>\%
  tamp()
}
\seealso{
Other cubble verbs include \code{\link{stretch}} and \code{\link{migrate}}
}
