% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsibble.R
\name{fill_gaps.temporal_cubble_df}
\alias{fill_gaps.temporal_cubble_df}
\alias{scan_gaps.temporal_cubble_df}
\title{\code{tsibble} functions implemented in cubble}
\usage{
\method{fill_gaps}{temporal_cubble_df}(.data, ..., .full = FALSE, .start = NULL, .end = NULL)

\method{scan_gaps}{temporal_cubble_df}(.data, ...)
}
\arguments{
\item{.data, ..., .full, .start, .end}{see \code{tsibble} documentation}
}
\value{
a cubble object
}
\description{
\code{tsibble} functions implemented in cubble
}
\examples{
library(tsibble)
climate_aus \%>\% face_temporal() \%>\% fill_gaps()
climate_aus \%>\% face_temporal() \%>\% scan_gaps()
}
